Refactoring Fix performance for get relationships with candidates
